requirements.txt
ipykernel
langchain
langgraph
langchain-core
langchain-community
langchain-openai
openai 
langchain-anthropic
anthropic
python-dotenv
rich
langchain-groq
groq
pydantic 
arxiv
wikipedia
langgraph-cli[inmem]